generator/requirements.txt
psycopg2-binary==2.9.7
prometheus_client==0.17.1